Se mueven algunos métodos de app.js a rutas y controladores
